+++ examples/multipass_examples/02-gdrive-check-user-credentials.py
@@ -8,8 +8,7 @@ from typing import List
 
 from google.oauth2.credentials import Credentials
 from llama_index.readers.google import GoogleDriveReader
-from pangea_multipass import (GDriveAPI, GDriveME, PangeaMetadataKeys,
-                              enrich_metadata)
+from pangea_multipass import (GDriveAPI, GDriveME, PangeaMetadataKeys, enrich_metadata)
 from pangea_multipass_llama_index import LIDocumentReader
 
 # Suppress specific warning
@@ -67,10 +66,8 @@ def google_drive_read_docs() -> List:
 documents = google_drive_read_docs()
 
 # Inference
-
 from pangea_multipass import GDriveAPI
-from pangea_multipass_llama_index import (LlamaIndexGDriveProcessor,
-                                          NodePostprocessorMixer)
+from pangea_multipass_llama_index import (LlamaIndexGDriveProcessor, NodePostprocessorMixer)
 
 # Create GDrive filter
 credentials_filepath = os.path.abspath("../credentials.json")