+++ packages/pangea-multipass-langchain/pangea_multipass_langchain/langchain.py
@@ -72,10 +72,11 @@ class LangChainGDriveFilter(GDriveProcessor[Document]):
 
     Args:
         creds (Credentials): Google OAuth2 credentials.
+        user_email (Optional[str]): User email to check access to files.
     """
 
-    def __init__(self, creds: Credentials):
-        super().__init__(creds, get_node_metadata=get_doc_metadata)
+    def __init__(self, creds: Credentials, user_email: Optional[str] = None):
+        super().__init__(creds, get_node_metadata=get_doc_metadata, user_email=user_email)
 
 
 class LangChainGithubFilter(GithubProcessor[Document]):