+++ packages/pangea-multipass/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Github repository's reader and processor.
 - Slack channel's reader and processor.
 - `account_id` support on JiraProcessor.
+- Check user email permissions with admin credentials in GDriveProcessor.
 
 ## 0.1.0 - 2024-12-24